Update the delete content item service with the changes to the delete content
item helper.

+++ app/services/delete_content_item.rb
@@ -20,9 +20,11 @@ module Services
         Translation,
         Unpublishing,
         UserFacingVersion,
+        ChangeNote,
       ]
 
       supporting_classes.each do |klass|
+        next unless ActiveRecord::Base.connection.data_source_exists?(klass.table_name)
         klass.where(content_item: content_items).destroy_all
       end
